fix: remove visual selection state when dragging cards
Deselection logic was moved to OnBeginDrag to ensure the card's visual state resets before being played. This prevents selected cards from remaining offset in the VisualCardsHandler after being played.
