--- Assets/Scripts/GameReactions/ClearBoardReactionsPRE.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c36d414276430834291f0bab932a3ec1
