--- Assets/Scripts/GameReactions/GenerateDecksReactionsPOST.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a7c2e77d4fb047c4bac02df01acf4d9c
