--- Assets/Scripts/GameReactions/PlayCardReactionsPOST.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: bbb4f2be5e4955c4b8523afa1258e189
