--- Assets/Scripts/GameReactions/ResolveRoundReactionsPOST.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: bf39ae487b581a645b57d762626cb645
