--- Assets/Scripts/GameReactions/StartGameReactionsPOST.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 85ccc96f4df4a7841b8f717379e1c6a2
