+++ Assets/Sprites/Interface/Tabuleiro Background.png.meta
@@ -147,7 +147,8 @@ TextureImporter:
     secondaryTextures: []
     spriteCustomMetadata:
       entries: []
-    nameFileIdTable: {}
+    nameFileIdTable:
+      Tabuleiro Background_0: -8462765999735424977
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
